move react table cell ui directory
